AI/gguf/challenge/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
gguf==0.9.1
llama-cpp-python==0.2.70
